refactor(oma-pm): do not connect zbus session in `OmaApt::new`
