+++ oma-pm/src/commit.rs
@@ -11,10 +11,11 @@ use oma_fetch::{Event, Summary, reqwest::Client};
 use oma_pm_operation_type::{InstallEntry, OmaOperation};
 use std::io::Write;
 use tracing::debug;
+use zbus::Connection;
 
 use crate::{
     apt::{DownloadConfig, OmaApt, OmaAptError, OmaAptResult},
-    dbus::change_status,
+    dbus::{OmaBus, Status, change_status},
     download::download_pkgs,
     progress::{InstallProgressArgs, InstallProgressManager, OmaAptInstallProgress},
 };
@@ -31,6 +32,24 @@ pub struct DoInstall<'a> {
     client: &'a Client,
     sysroot: &'a str,
     config: CommitNetworkConfig<'a>,
+    conn: Option<Connection>,
+}
+
+async fn create_session() -> Result<Connection, zbus::Error> {
+    let conn = zbus::connection::Builder::system()?
+        .name("io.aosc.Oma")?
+        .serve_at(
+            "/io/aosc/Oma",
+            OmaBus {
+                status: Status::Pending,
+            },
+        )?
+        .build()
+        .await?;
+
+    debug!("zbus session created");
+
+    Ok(conn)
 }
 
 impl<'a> DoInstall<'a> {
@@ -40,11 +59,22 @@ impl<'a> DoInstall<'a> {
         sysroot: &'a str,
         config: CommitNetworkConfig<'a>,
     ) -> Result<Self, OmaAptError> {
+        let conn = apt.tokio.block_on(async {
+            match create_session().await {
+                Ok(conn) => Some(conn),
+                Err(e) => {
+                    debug!("Failed to create D-Bus session: {:?}", e);
+                    None
+                }
+            }
+        });
+
         Ok(Self {
             apt,
             sysroot,
             client,
             config,
+            conn,
         })
     }
 
@@ -78,7 +108,7 @@ impl<'a> DoInstall<'a> {
             .map_err(|e| OmaAptError::FailedOperateDirOrFile(path.display().to_string(), e))?;
 
         self.apt.tokio.block_on(async {
-            if let Some(conn) = &self.apt.conn {
+            if let Some(conn) = &self.conn {
                 change_status(conn, "Downloading").await.ok();
             }
 
@@ -114,7 +144,7 @@ impl<'a> DoInstall<'a> {
         let args = InstallProgressArgs {
             config: self.apt.config,
             tokio: self.apt.tokio,
-            connection: self.apt.conn,
+            connection: self.conn,
         };
 
         let mut progress =
